Fix the resteasy-reactor-netty tests after JUnit 5.11 upgrade.
